requirements.txt
strands-agents[anthropic]>=0.1.0
Flask>=2.3.0
flask-cors>=3.0.10 